Add files via upload
This is the first assignment on Logistic Regression Classifier to recognize letters from 'A' to 'J'.

The data is rough and more than half of the time, the code was to clean it up into numpy.array and to be used by Logistic Regression training package. The process of getting and cleaning the data is very helpful and was not taught by Machine learning course from Coursera.

Using 5000 samples, I got an accuracy 85% on the test set.